+++ platform/src/main/java/com/zzz/platform/api/invoice/service/InvoiceUploadService.java
@@ -4,7 +4,9 @@ import com.zzz.platform.api.invoice.domain.InvoiceJsonVO;
 import com.zzz.platform.common.domain.ResponseDTO;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigInteger;
+
 public interface InvoiceUploadService {
 
-    ResponseDTO<InvoiceJsonVO> upload(Long userId, MultipartFile file);
+    ResponseDTO<InvoiceJsonVO> upload(BigInteger userId, MultipartFile file);
 }
